Create initial docs (#4)
* init docs by sphinx.

* Update documentation theme to blue

* add doc to README
