+++ docs/source/conf.py
@@ -11,11 +11,12 @@
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
-#
+
 # import os
 # import sys
-# sys.path.insert(0, os.path.abspath('.'))
+# sys.path.insert(0, os.path.abspath('..'))
 import sphinx_rtd_theme
+import yews
 
 
 # -- Project information -----------------------------------------------------
@@ -25,7 +26,7 @@ copyright = '2019, Lijun Zhu'
 author = 'Lijun Zhu'
 
 # The short X.Y version
-version = 'master (' + '0.0.1' + ')'
+version = 'master (' + yews.__version__ + ')'
 # The full version, including alpha/beta/rc tags
 release = 'master'
