+++ docs/source/models.rst
@@ -1,0 +1,15 @@
+yews.models
+===========
+
+The models subpackage contains definitions for the following model
+architectures:
+
+.. contents:: Models
+   :local:
+
+.. currentmodule:: yews.models
+
+CPIC
+----
+
+.. autofunction:: cpic64
