+++ docs/source/train.rst
@@ -1,0 +1,8 @@
+yews.train
+----------
+
+.. currentmodule:: yews.train
+
+.. autoclass:: Trainer
+   :members: train, validate
+   :special-members:
