--- yews/dataset/__init__.py
@@ -1,10 +1,0 @@
-from .classification import ClassificationDataset
-from .array import ClassificationDatasetArray
-from .folder import ClassificationDatasetFolder
-
-__all__ = (
-    'ClassificationDataset',
-    'ClassificationDatasetArray',
-    'ClassificationDatasetFolder',
-)
-
