+++ yews/datasets/__init__.py
@@ -1,0 +1,8 @@
+from .classification import ClassificationDataset, DatasetArray, DatasetFolder
+
+__all__ = (
+    'ClassificationDataset',
+    'DatasetArray',
+    'DatasetFolder',
+)
+
